6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,14 +1,17 @@
 #include <stdio.h>
+
 /**
  * main -  program that prints the size of various types on the computer it is compiled and run on
  * return o
  */
+
 int main(void)
+
 {
 	char a;
 	int b;
 	long int c;
-	long long int d;
+	long int d;
 	float f;
 
 printf("size of the char: %lu byte(s)\n", (unsiggned long)sizeof(a));
